# perf(engine): reuse existing array in genericFunction.Evaluate